Memory copying is really slow. Should try to copy weights and biases in advance.

+++ c_implementation/weights/slike.h
@@ -1,6 +1,6 @@
 //Testiranje ispisa
 //Niz je: (784,)
-extern double sedmica[784];
+extern float sedmica[784];
 extern double dvojka[784];
 extern double trojka[784];
 extern double petica[784];
